Assignment - 1 - Database Design/requirements.txt
kaggle
pandas
numpy
mysql-connector-python